+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.0.0-beta.1
 
+- Feat: ensure uniform usage of 'volgnummers' across different clients
 - Feat: enable filtering on orderBy in gutenberg block
 - Feat: manual configuration of form settings, besides selecting retrieved settings from external source
 - Refactor: general improvements and cleaning-up